requirements/requirements.txt
appnope==0.1.2
asgiref==3.3.4
backcall==0.2.0
bcrypt==3.2.0
boto3==1.17.103
botocore==1.20.103
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.45.0
django-cors-headers==3.7.0
django-extensions==3.1.3
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
matplotlib-inline==0.1.2
mock==4.0.3
mysqlclient==2.0.3
oauthlib==3.1.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
PyMySQL==1.0.2
pyproj==3.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5